tests/wpt/tests/html/semantics/embedded-content/media-elements/volume_nonfinite.html

<!doctype html>
<meta charset="utf-8">
<title>Setting HTMLMediaElement.volume to non-finite numbers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
["audio", "video"].forEach(function(aElement) {
  [NaN, Infinity, -Infinity].forEach(function(aValue) {
    test(function() {
      var el = document.createElement(aElement);
      assert_throws_js(TypeError, function() {
        el.volume = aValue;
      });
    }, "Setting " + aElement + ".volume to " + String(aValue) + " should throw a TypeError");
  });
});
</script>
